+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.13.1 / 2020-08-08
+===================
+
+* Fixes Dockerfile
+* Error messages improved for timeout errors
+* Updates dependencies
+
 2.13.0 / 2020-07-24
 ===================